fix s3 test

+++ tests/docker_compatible_gcs/run.sh
@@ -38,6 +38,11 @@ export GOOGLE_APPLICATION_CREDENTIALS="tests/$TEST_NAME/config.json"
 for TAG in ${TAGS}; do
     echo "restore ${TAG} data starts..."
     bin/br restore db --db test -s "gcs://$BUCKET/bk${TAG}" --pd $PD_ADDR --gcs.endpoint="http://$GCS_HOST:$GCS_PORT/storage/v1/"
+    row_count=$(run_sql_in_container  "SELECT COUNT(*) FROM test.usertable;" | awk '/COUNT/{print $2}')
+    if [ $row_count != $EXPECTED_KVS ]; then
+       echo "restore kv count is not as expected(1000) $row_count"
+       exit 1
+    fi
     # clean up data for next restoration
     run_sql_in_container "drop database test;"
 done
